♻️ unify exec interface
