+++ impl-c/include/proc/exec.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void *load_program(const char *name);
+void exec(const char *name, const char **argv);
